dockers/llm.rag.service/requirements.txt
faiss-cpu
fastapi
langchain-huggingface
uvicorn
opentelemetry-api
opentelemetry-instrumentation
opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-http
arize-phoenix
arize-phoenix[evals]
tiktoken
openinference-instrumentation-langchain
nest-asyncio 
langchain
langchain_community
langchain-community
opentelemetry-sdk
opentelemetry-exporter-otlp
openai>=1
langchain-openai
httpx<0.28
pandas
sqlalchemy
weaviate
